[Baekjoon(JAVA)/수학] 13171_A

+++ Baekjoon/Main_13171_A.java
@@ -4,25 +4,20 @@ import java.io.InputStreamReader;
 
 public class Main_13171_A {
 
-	private static int A;
-	private static int X;
+	private static long A;
+	private static long X;
 	private static long[] arr;
 
 	public static void main(String[] args) throws NumberFormatException, IOException {
 		BufferedReader bf = new BufferedReader(new InputStreamReader(System.in));
-		A = Integer.parseInt(bf.readLine());
-		X = Integer.parseInt(bf.readLine());
+		A = Long.parseLong(bf.readLine());
+		X = Long.parseLong(bf.readLine());
 		arr = new long[65];
 		
 		arr[0] = A % 1000000007;
 		for (int i = 1; i < 64; i++) {
 			arr[i] = ((arr[i-1] * arr[i-1]) % 1000000007); 
 		}
-		
-//		for (int i = 0; i < 64; i++) {
-//			System.out.print(arr[i] + " ");
-//		}
-//		System.out.println();
 		
 		long ans = 1;
 		int count = 0;
